Loads of stuff

+++ src/main/java/uk/co/idinetwork/core/controller/management/SiteNavigationController.java
@@ -1,5 +1,7 @@
 package uk.co.idinetwork.core.controller.management;
 
+import javax.servlet.http.HttpServletRequest;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -8,6 +10,7 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.servlet.ModelAndView;
 
 import uk.co.idinetwork.core.model.SiteNavigation;
+import uk.co.idinetwork.core.service.AuthorisationService;
 import uk.co.idinetwork.core.service.SiteNavigationService;
 import uk.co.idinetwork.core.utils.KeyUtil;
 
@@ -18,39 +21,58 @@ public class SiteNavigationController {
 	private static final String CONFIRMATION_VIEW = "management/site-navigation-confirmation";
 	
 	@Autowired private SiteNavigationService siteNavigationService;
+	@Autowired private AuthorisationService authorisationService;
 
 	@RequestMapping(value=CONTROLLER_MAPPING, method=RequestMethod.GET)
-	public ModelAndView siteNavigationList() {
+	public ModelAndView siteNavigationList(HttpServletRequest request) {
 		ModelAndView modelAndView = new ModelAndView(VIEW);
 		
-		modelAndView.addObject("topSiteNavigation", siteNavigationService.findTopNavigation());
-		modelAndView.addObject("sideSiteNavigation", siteNavigationService.findSideNavigation());
+		if (authorisationService.isAuthorised()) {
+			modelAndView.addObject("topSiteNavigation", siteNavigationService.findTopNavigation());
+			modelAndView.addObject("sideSiteNavigation", siteNavigationService.findSideNavigation());
+		}
+		else {
+        	modelAndView = new ModelAndView(authorisationService.getView());
+        	modelAndView.addObject("redirectUrl", authorisationService.getLoginUrl(request.getRequestURI()));
+		}
 		
 		return modelAndView;
 	}
 	
 	@RequestMapping(value=CONTROLLER_MAPPING + "/add/{type}", method=RequestMethod.POST)
-	public ModelAndView addTopNavigation(@PathVariable ("type") String type, String title) {
+	public ModelAndView addNavigation(HttpServletRequest request, @PathVariable ("type") String type, String title) {
 		ModelAndView modelAndView = new ModelAndView(CONFIRMATION_VIEW);
 		
-		SiteNavigation siteNavigation = new SiteNavigation();
-		siteNavigation.setKey(KeyUtil.buildKey(title));
-		siteNavigation.setTitle(title);
-		siteNavigation.setType(type);
-		siteNavigationService.saveSiteNavigation(siteNavigation);
-		
-		if (siteNavigation.getId() != null) {
-			modelAndView.addObject("siteNavigationConfirmation", true);
+		if (authorisationService.isAuthorised()) {
+			SiteNavigation siteNavigation = new SiteNavigation();
+			siteNavigation.setKey(KeyUtil.buildKey(title));
+			siteNavigation.setTitle(title);
+			siteNavigation.setType(type);
+			siteNavigationService.saveSiteNavigation(siteNavigation);
+			
+			if (siteNavigation.getId() != null) {
+				modelAndView.addObject("siteNavigationConfirmation", true);
+			}
+		}
+		else {
+        	modelAndView = new ModelAndView(authorisationService.getView());
+        	modelAndView.addObject("redirectUrl", authorisationService.getLoginUrl(request.getRequestURI()));
 		}
 		
 		return modelAndView;
 	}
 	
 	@RequestMapping(value=CONTROLLER_MAPPING + "/delete/{type}/{id}", method=RequestMethod.POST)
-	public ModelAndView deleteTopNavigation(@PathVariable ("type") String type, @PathVariable ("id") Long id) {
+	public ModelAndView deleteTopNavigation(HttpServletRequest request, @PathVariable ("type") String type, @PathVariable ("id") Long id) {
 		ModelAndView modelAndView = new ModelAndView(CONFIRMATION_VIEW);
 		
-		modelAndView.addObject("siteNavigationConfirmation", siteNavigationService.deleteSiteNavigation(type, id));
+		if (authorisationService.isAuthorised()) {
+			modelAndView.addObject("siteNavigationConfirmation", siteNavigationService.deleteSiteNavigation(type, id));
+		}
+		else {
+        	modelAndView = new ModelAndView(authorisationService.getView());
+        	modelAndView.addObject("redirectUrl", authorisationService.getLoginUrl(request.getRequestURI()));
+		}
 		
 		return modelAndView;
 	}
